requirements.txt
streamlit
numpy
pandas
scikit-learn
xgboost
joblib
matplotlib
seaborn
Pillow
